+++ aphrodite/v1/worker/gpu/sample/min_p.py
@@ -14,7 +14,7 @@ def _min_p_kernel(
     vocab_size,
     BLOCK_SIZE: tl.constexpr,
 ):
-    token_idx = tl.program_id(0)
+    token_idx = tl.program_id(0).to(tl.int64)
     req_state_idx = tl.load(expanded_idx_mapping_ptr + token_idx)
     min_p = tl.load(min_p_ptr + req_state_idx).to(tl.float32)
     if min_p == 0.0:
@@ -45,7 +45,9 @@ def _min_p_kernel(
         tl.store(logits_ptr + token_idx * logits_stride + block, logits, mask=mask)
 
 
-def apply_min_p(logits: torch.Tensor, expanded_idx_mapping: torch.Tensor, min_p: torch.Tensor) -> None:
+def apply_min_p(
+    logits: torch.Tensor, expanded_idx_mapping: torch.Tensor, min_p: torch.Tensor
+) -> None:
     num_tokens, vocab_size = logits.shape
     BLOCK_SIZE = 1024
     _min_p_kernel[(num_tokens,)](
